requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29